chore: refactoring component location
