+++ frontend/src/components/NoteComponent.vue
